cleanup and refactor side wide shared JS code
