+++ source/js/common/index.js
@@ -1,0 +1,7 @@
+// The following modules are the common modules which need to be called
+// on all pages on the main Foundation site, apps, and microsites.
+// e.g., Foundation site, PNI, MozFest etc
+
+export { default as bindEventHandlers } from "./bind-event-handlers";
+export { default as initiatePrimaryNav } from "./initiate-primary-nav";
+export { default as injectReactComponents } from "./inject-react-components";
