+++ source/js/main.js
@@ -5,8 +5,8 @@ import ReactGA from "react-ga";
 import ReactDOM from "react-dom";
 import Analytics from "./analytics.js";
 import * as Sentry from "@sentry/browser";
+import * as common from "./common";
 
-import JoinUs from "./components/join/join.jsx";
 import Petition from "./components/petition/petition.jsx";
 import MultipageNavMobile from "./components/multipage-nav-mobile/multipage-nav-mobile.jsx";
 import News from "./components/news/news.jsx";
@@ -14,7 +14,6 @@ import PulseProjectList from "./components/pulse-project-list/pulse-project-list
 import ShareButtonGroup from "./components/share-button-group/share-button-group.jsx";
 
 import primaryNav from "./primary-nav.js";
-import navNewsletter from "./nav-newsletter.js";
 import bindMozFestGA from "./mozfest-ga.js";
 import bindMozFestEventHandlers from "./mozfest-event-handlers";
 import youTubeRegretsTunnel from "./youtube-regrets.js";
@@ -26,7 +25,7 @@ if (
   process.env.NODE_ENV === "development"
 ) {
   axe = require("react-axe");
-  axe(React, ReactDOM, 1000);
+  // axe(React, ReactDOM, 1000);
 }
 
 // To be populated via XHR and querySelector
@@ -205,8 +204,7 @@ let main = {
     // Call once to get scroll position on initial page load.
     onScroll();
 
-    primaryNav.init();
-    navNewsletter.init(networkSiteURL, csrfToken);
+    common.initiatePrimaryNav(networkSiteURL, csrfToken, primaryNav);
     youTubeRegretsTunnel.init();
 
     // Extra tracking
@@ -222,16 +220,7 @@ let main = {
       });
     }
 
-    let donateFooterBtn = document.getElementById(`donate-footer-btn`);
-    if (donateFooterBtn) {
-      donateFooterBtn.addEventListener(`click`, () => {
-        ReactGA.event({
-          category: `donate`,
-          action: `donate button tap`,
-          label: `${document.title} footer`
-        });
-      });
-    }
+    common.bindEventHandlers();
   },
 
   bindGAEventTrackers() {
@@ -271,25 +260,7 @@ let main = {
 
   // Embed various React components based on the existence of containers within the current page
   injectReactComponents() {
-    // Embed additional instances of the Join Us box that don't need an API exposed (eg: Homepage)
-    document.querySelectorAll(`.join-us:not(.on-nav)`).forEach(element => {
-      var props = element.dataset;
-
-      props.apiUrl = `${networkSiteURL}/api/campaign/signups/${props.signupId ||
-        0}/`;
-
-      props.csrfToken = props.csrfToken || csrfToken;
-      props.isHidden = false;
-
-      apps.push(
-        new Promise(resolve => {
-          ReactDOM.render(
-            <JoinUs {...props} whenLoaded={() => resolve()} />,
-            element
-          );
-        })
-      );
-    });
+    common.injectReactComponents(apps, networkSiteURL, csrfToken);
 
     // petition elements
     var subscribed = false;
